Revert "Use TS for the Jest config file"
This reverts commit eed19291235fe57d4e6c5c4151151094be8fa2c1.
