+++ package.json
@@ -124,7 +124,6 @@
     "stylelint": "^14.8.0",
     "stylelint-config-standard": "^25.0.0",
     "to-vfile": "7.2.3",
-    "ts-node": "^10.9.1",
     "tsc-esm-fix": "^2.13.0",
     "tsm": "^2.2.1",
     "typescript": "^4.6.3",
